chore: rename blend node
