+++ packages/effects-core/src/plugins/animation-graph/nodes/blend-node.ts
